Move open_intent_envelope to intent_service
